requirements.txt
certifi==2020.12.5
grpc-testing==0.0.1
grpcio==1.38.0
grpcio-testing==1.38.0
grpcio-tools==1.38.0
protobuf==3.17.1
six==1.16.0
